Looking for just one thing possible now
Usually, if you have a sparql query with only optional statements and a filter on a variable ?var0, then the results are empty.
The SPEX tool is supposed to work even then! So for this purpose, in a situation as described, following pattern is added to the query:
{?var0?x_var0?y_var0.}union{?x_var0?y_var0?var0.}
Because the header says "SELECT DISTINCT * WHERE", multiple results are shown. So, at the moment, this is solved by setting the Limit of the query to 1.

These changes to the query are not visible in the query window. To know why, look at "this.showquery" in spex.js.