Single click enough to release selection now
If an item in the results-pane is selected, it is highlighted and zoomed on in the map. This can be done by clicking, or mouseover. Earlier, to release a selected item by click, one had to double-click. Now, a single-click is enough. Also, the hint "Click to release selection!" appears on hover. This doesn't look fancy like the other tooltips yet, could somebody else do this who knows how? ;)